Figure 4.  Spike region TTMV-Ly1 protomer with P1 and P2 domains forming a 5-fold symmetry crown capsid structure
The spike region consists of the inner P1 and outermost P2 domain (A).  Tip of the P2 domain faces interstitial space and is accessible to antibodies.  Also, a 5-fold symmetry crown structure of capsid protomers may play a role in cellular interaction (B).  Notably, C and N terminus are missing in these TTMV-Ly1 protomers used to Cryo-EM study.  In particular, a portion of the C terminus is also exposed to the interstitial space, and my well participate in the cell interaction.