Sepplemental Figure 1. Alignment of 6 known Tregitopes  against TTV1-ORF1 (XCH55654.1)
Pairwise alignment of those previously characterized 6 Tregitopes (33) was done by using R package, DECIPHER (34)and Biostrings (35).  The top sequence is the primary sequence of TTV1-ORF1 and following sequences 2 through 11 are the peptide sequences of known Tregitopes.  Numbers on the top of each rows displays amino acid position starting from 1st amino acid position in ORF1, and the bottom sequence denoted as consensus shows the amino acids in agreement between Tregitopes and TTV1-ORF1 primary sequences.  Notice the “x” in the consensus sequence which indicates that the amino acids in Tregitopes and TTV1-ORF are not in agreement.  AlignPair function in DECIPHER was used with a default parameter.   A black dot represents a  matched amino acid between the primary sequence of Tregitopes and TTV1-ORF1.